+++ scout/server/blueprints/genes/views.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
-from flask import abort, Blueprint, request, jsonify, redirect, url_for, flash
+from flask import Blueprint, abort, flash, jsonify, redirect, request, url_for
 
 from scout.server.extensions import store
-from scout.server.utils import templated, public_endpoint
+from scout.server.utils import public_endpoint, templated
+
 from . import controllers
 
 genes_bp = Blueprint("genes", __name__, template_folder="templates")